SPSAOptimizers

Documentation for SPSAOptimizers.

SPSAOptimizers.DictStreams

SPSAOptimizers.PowerStreams

SPSAOptimizers.QiskitInterface

SPSAOptimizers.RandomStreams

SPSAOptimizers.Serialization

SPSAOptimizers.ConstantStreams.ConstantSeries

SPSAOptimizers.DictStreams.IntDictStream

SPSAOptimizers.FirstOrderOptimizers.FirstOrderOptimizer

SPSAOptimizers.Hessians.HessianType

SPSAOptimizers.Optimizers.OptimizerType

SPSAOptimizers.PowerStreams.PowerSeries

SPSAOptimizers.RandomStreams.BernoulliDistribution

SPSAOptimizers.SPSA1s.SPSA1

SPSAOptimizers.SPSA2s.SPSA2

SPSAOptimizers.SecondOrderOptimizers.SecondOrderOptimizer

SPSAOptimizers.Streams.StreamType

SPSAOptimizers.TrajectoryHessians.TrajectoryHessian

SPSAOptimizers.Hessians.regularized

SPSAOptimizers.Hessians.update!

SPSAOptimizers.Optimizers.Record

SPSAOptimizers.Optimizers.Trace

SPSAOptimizers.Optimizers.averagerecords!

SPSAOptimizers.Optimizers.copyrecord!

SPSAOptimizers.Optimizers.iterate!

SPSAOptimizers.Optimizers.optimize!

SPSAOptimizers.Optimizers.trace!

SPSAOptimizers.Optimizers.trajectory

SPSAOptimizers.QiskitInterface.SPSA

SPSAOptimizers.QiskitInterface.minimize

SPSAOptimizers.RandomStreams.sample

SPSAOptimizers.Serialization.__data__

SPSAOptimizers.Serialization.deserialize

SPSAOptimizers.Serialization.init

SPSAOptimizers.Serialization.load!

SPSAOptimizers.Serialization.reset!

SPSAOptimizers.Serialization.serialize

SPSAOptimizers.Streams.next!

SPSAOptimizers.Streams.next!

SPSAOptimizers.Serialization — Module

SPSAOptimizers.Serialization.__data__ — Function

SPSAOptimizers.Serialization.deserialize — Method

deserialize(json)

Convert a JSON-serializable NamedTuple into an object.

SPSAOptimizers.Serialization.init — Function

init(::Type{<:Serializable}, data)

Ignore mutable attributes in data; just set state-independent values.

SPSAOptimizers.Serialization.load! — Function

load!(::Serializable, data)

Ignore immutable attributes in data; just change state-dependent values.

SPSAOptimizers.Serialization.reset! — Function

reset!(::Serializable)

Restore mutable attributes to their initial values.

SPSAOptimizers.Serialization.serialize — Method

serialize(object)

Convert an object into a JSON-serializable NamedTuple.

SPSAOptimizers.Streams.StreamType — Type

StreamType{T}

Concrete StreamTypes should implement the interface for serialization (data, init, load!, reset!, along with calling register) in addition to the next! function, which returns an object of type T.

SPSAOptimizers.Streams.next! — Function

next!(stream::StreamType{T})

Returns the next element of type T from the given stream object.

SPSAOptimizers.PowerStreams — Module

Guidelines for selecting gain sequences:

Let η[k] = η0 / (A + k + 1)^α h[k] = h0 / (k + 1)^γ

Asymptotically optimal: α=1, γ=1/6

Lowest allowable for theoretical convergence: α=0.602, γ=0.101

Empirically, the smaller the better, but maybe for large problems one should transition to the larger.

Set h0 to the standard deviation of the noise in f.

Set A to 10% of the maximum number of iterations.

Set a0 such that a[0] * |g(x0)| gives a desirable "change in magnitude" of x.

SPSAOptimizers.PowerStreams.PowerSeries — Type

PowerSeries(a0, γ, A)

Yield from the sequence $a[k] = a_0 / (A + k + 1)^γ$.

SPSAOptimizers.ConstantStreams.ConstantSeries — Type

ConstantSeries(C)

Yield the number C, every time.

SPSAOptimizers.Streams.next! — Method

SPSAOptimizers.RandomStreams — Module

SPSAOptimizers.RandomStreams.BernoulliDistribution — Type

BernoulliDistribution(L, k, p, seed)

A Bernoulli distribution yielding a vector of +/-1.

Parameters

L::Int, the length of the vector to generate

k::Int=L, the number of elements in the vector to randomize If an element isn't randomized, it is set to 0. Which elements to randomize are...selected randomly. :)

p::Int=1.0, the probability of randomizing any given element in the vector

NOTE: k and p are different and mutually exclusive ways to let 0's sneak into the Bernoulli distribution. The defaults are just "no 0's allowed", i.e. an actual Bernoulli distribution.

seed::Int, the random seed defining the pseudorandom stream

SPSAOptimizers.RandomStreams.sample — Method

NOTE: r=1.0 is defined to give 0, irrespective of X.p

SPSAOptimizers.DictStreams — Module

SPSAOptimizers.DictStreams.IntDictStream — Type

IntDictStream(dict, default)

A customized stream specifying the output for any given specific iteration in dict, or otherwise a fixed constant default.

SPSAOptimizers.Hessians.HessianType — Type

An internal representation of a "tracked" Hessian, e.g. the average of Hessians measured at each iteration in the trajectory.

SPSAOptimizers.Hessians.regularized — Function

regularized(H::HessianType{F})

A bit of post-processing to construct a positive semi-definite version of H.

SPSAOptimizers.Hessians.update! — Function

update!(H::HessianType{F}, Hm::AbstractMatrix{F})

Update the "tracked" Hessian H with the "measured" Hessian Hm.

SPSAOptimizers.TrajectoryHessians.TrajectoryHessian — Type

TrajectoryHessian(bias, H, k)

Hessian is average of Hessians from each iteration in trajectory.

Parameters

bias::Float - the regularization bias, i.e. the iotum of an identity matrix to add when computing a positive semi-definite version of this matrix

H::Matrix{Float} - the average Hessian over all k iterations

k::Ref{Int} - the number of iterations over which H is averaged

TrajectoryHessian(L; bias)

Convenience constructor which initializes H to an LxL matrix, and k to 0. The bias defaults to 0.01 but can be overridden.

Note that we initialize H to identity, but this does not matter because, with k=0, it will be weighted by 0 in the first update!

SPSAOptimizers.Optimizers.OptimizerType — Type

Used for dispatch to different optimization algorithms.

Algorithms are organized by first-order, second-order, and so on. (Well, just those two for now, but I feel like there could be a "quasi-second order" someday.)

Constructors are documented with concrete types (e.g. SPSA1), but details on optimization options, record schematics, and iterate! keyword arguments are found in the intermediate order types, e.g. FirstOrderOptimizer.

SPSAOptimizers.Optimizers.Record — Function

Record(optimizer, L)

Initialize a mutable object giving the status of an L-dimensional optimization.

The object is a NamedTuple of vectors and references, whose precise schema depends on the type of optimizer.

SPSAOptimizers.Optimizers.Trace — Method

Trace()

Initialize a vector of Records, giving a trace of an optimization.

Simply allows for the more intuitive trace initialization, Trace(). Asking people to type NamedTuple[] just sounds so...pretentious.

SPSAOptimizers.Optimizers.averagerecords! — Method

averagerecords!(archive, records)

Like copyrecord!, but stores an average of many records in archive.

SPSAOptimizers.Optimizers.copyrecord! — Method

copyrecord!(archive, record)

Copy all values from the record Record to the archive Record.

SPSAOptimizers.Optimizers.iterate! — Function

iterate!(optimizer, fn, x; kwargs...)

Advance by one step in an optimization routine, updating x in place.

In addition to updating x, this will update the state of optimizer.

Parameters

optimizer::OptimizerType{F} - defines the optimization algorithm

fn - the loss-function, such that fn(x) returns a number of type F

x::AbstractVector{F} - the best guess so far

Keyword arguments depend on the type of optimizer.

SPSAOptimizers.Optimizers.optimize! — Function

optimize!(optimizer, fn, x0; kwargs...)

Run an optimization routine to convergence, updating x in place.

In addition to updating x, this will update the state of optimizer.

Parameters

optimizer::OptimizerType{F} - defines the optimization algorithm

fn - the loss-function, such that fn(x) returns a number of type F

x::AbstractVector{F} - the best guess so far

Keyword Arguments

maxiter::Int - the max number of iterations to attempt, successful or not

callback - function called at each iteration, successful or not:

callback(optimizer, iterate)

where iterate is a record of the proposed step. The callback is called immediately prior to deciding whether to accept or reject the step, and the step is always rejected if the callback returns true.

record, average_last, trace, tracefields - see Output section below

Additional keyword arguments may be available depending on the type of optimizer.

Output

A record (a NamedTuple with schema specified by the optimizer) is returned at the end. If you provide an integer to average_last, that record is an average of the last so many (successful) iterations.

If you provide an object to the record keyword argument, that object will be the one used for the final return output. More importantly, this same object is used to keep track of the last successful iteration, so it will have meaningful data if something (e.g. keyboard interrupt) disrupts the algorithm. Note however that if terminated prematurely, record will only represent the last successful iteration, rather than an average, even if average_last is provided.

You can also provide a trace, which is a list of records of every single step. Provide a collection of field names (as symbols) to specify which fields should be saved in the trace. The default is the collection of all scalar attributes. Set it to an empty collection to trace every possible attribute.

SPSAOptimizers.Optimizers.trace! — Method

trace!(trace, record, fields...)

Copies a record and adds the copy to the trace.

Optionally, list specific fields - all others will be omitted from the copy.

SPSAOptimizers.Optimizers.trajectory — Method

trajectory(trace, field)

Extract the values of a particular field from each record in a trace, and return as a vector (if field names a reference) or a matrix, where each column is the vector stored in one record.

SPSAOptimizers.FirstOrderOptimizers.FirstOrderOptimizer — Type

FirstOrderOptimizer{F}

Super-type for first-order optimizers. (I'm pretty sure there's just the one, SPSA1.) Constructors are documented with the concrete type(s?), but find details on optimization options, record schematics, and iterate! keyword arguments here:

Record Schematic

Record(::FirstOrderOptimizer{F}, L::Int)

Constructs a NamedTuple with fields

x::Vector{F} of length L, the current parameters

f::Ref{F}, the current function value

g::Vector{F} of length L, the current gradient estimate

p::Vector{F} of length L, the current parameter step (based on gradient)

xp::Vector{F} of length L, the proposed parameters for the next iteration

fp::Ref{F}, the function value at the next iteration

nfev::Ref{Int}, the number of function evaluations to date

time::Ref{F}, the algorithm time to date

bytes::Ref{F}, the memory consumption to date

Additional optimize! Keyword Arguments

trust::F=Inf, the maximum allowable norm of p

If p has a larger norm, it gets clipped to trust.

tolerance::F=Inf, the amount by which fp may exceed f without rejecting the step

iterate! Interface

iterate!(
    optimizer, fn, x;
    # EXTRA OPTIONS
    trust,
    # TRACEABLES
    f, g, p, xp, nfev,
    # WORK ARRAYS
    __xe
)

optimizer, fn, x: the mandatory positional arguments, see iterate!

trust: see above

f, g, p, xp, nfev: when provided, stores meaningful outputs (when called through optimize!, these are provided from the record of the last successful iteration).

__xe: a work variable matching the dimensions of x, used to store the stochastically perturbed parameters for finite difference

SPSAOptimizers.SPSA1s.SPSA1 — Type

SPSA1{P}(η, h, e, n)

The first-order SPSA optimizer.

Type Parameters

P::Int, the order of finite difference, typically 2

Parameters

η - the float stream for step-length at each iteration

h - the float stream for finite-difference perturbation at each iteration

e - the vector stream for each finite difference perturbation direction This also determines the dimensions in the parameter vector.

n - the int stream for number of times to sample the gradient at each iteration

SPSA1(L; P, η, h, e, n)

With this constructor, you need only provide the number of dimensions L, and all the above parameters will be filled in by sensible defaults (but you can override any of them with the kwarg).

Additionally, you may use the η and h kwargs to specify tuples defining a power series.

Defaults

P: 2, of course - central finite difference

η: a power series with a0 = 0.2, alpha=0.602, and A = 0

h: a power series with c0 = 0.2 and gamma=0.101

e: a fair coin-toss between +/-1 for each dimension

n: one sample for every iteration

Tuple Interface

If you don't feel like manually constructing the PowerStream object, you can just pass a tuple directly here.

η: (a0, alpha, A) all floats (even though A is semantically an integer)

h: (c0, gamma) all floats (note A is always 0 for the h stream)

SPSAOptimizers.SecondOrderOptimizers.SecondOrderOptimizer — Type

SecondOrderOptimizer{F}

Super-type for first-order optimizers. (Just SPSA2 for now but I have visions of how to enhance. They probably aren't worth the effort, though.) Constructors are documented with the concrete type(s?), but find details on optimization options, record schematics, and iterate! keyword arguments here:

Record Schematic

Record(::FirstOrderOptimizer{F}, L::Int)

Constructs a NamedTuple with fields

x::Vector{F} of length L, the current parameters

f::Ref{F}, the current function value

g::Vector{F} of length L, the current gradient estimate

H::Matrix{F} of size(L,L), the current Hessian estimate

p::Vector{F} of length L, the current parameter step (based on gradient)

xp::Vector{F} of length L, the proposed parameters for the next iteration

fp::Ref{F}, the function value at the next iteration

nfev::Ref{Int}, the number of function evaluations to date

time::Ref{F}, the algorithm time to date

bytes::Ref{F}, the memory consumption to date

Additional optimize! Keyword Arguments

trust::F=Inf, the maximum allowable norm of p

If p has a larger norm, it gets clipped to trust.

tolerance::F=Inf, the amount by which fp may exceed f without rejecting the step

warmup::Int=0, after how many iterations to begin refining the step with the Hessian (rather than just using gradient directly)

solver, a function with header solver(H, g) returning the "preconditioned" $H^{-1} g$

iterate! Interface

iterate!(
    optimizer, fn, x;
    # EXTRA OPTIONS
    trust, precondition, solver,
    # TRACEABLES
    f, g, H, p, xp, nfev,
    # WORK ARRAYS
    __xe,
    __e1,
)

optimizer, fn, x: the mandatory positional arguments, see iterate!

trust, solver: see above

precondition: whether or not to use the Hessian in this iteration

f, g, H, p, xp, nfev: when provided, stores meaningful outputs (when called through optimize!, these are provided from the record of the last successful iteration).

__xe: a work variable matching the dimensions of x, Used to store the stochastically perturbed parameters for finite difference

__e1: a work variable matching the dimensions of x Used to store one of the two perturbation vectors needed for second order finite difference. Note that the other one is stored within the stream object, which is why this isn't needed at all in first-order.

SPSAOptimizers.SPSA2s.SPSA2 — Type

SPSA2(H, η, h, e, n)

The second-order SPSA optimizer.

Parameters

H - the Hessian object, updated at each iteration

η - the float stream for step-length at each iteration

h - the float stream for finite-difference perturbation at each iteration

e - the vector stream for each finite difference perturbation direction This also determines the dimensions in the parameter vector.

n - the int stream for number of times to sample the gradient at each iteration

SPSA2(L; H, η, h, e, n)

With this constructor, you need only provide the number of dimensions L, and all the above parameters will be filled in by sensible defaults (but you can override any of them with the kwarg).

Additionally, you may use the η and h kwargs to specify tuples defining a power series.

Defaults

H: a fresh LxL TrajectoryHessian with regularization bias 0.01

η: a power series with a0 = 0.2, alpha=0.602, and A = 0

h: a power series with c0 = 0.2 and gamma=0.101

e: a fair coin-toss between +/-1 for each dimension

n: one sample for every iteration

Tuple Interface

If you don't feel like manually constructing the PowerStream object, you can just pass a tuple directly here.

η: (a0, alpha, A) all floats (even though A is semantically an integer)

h: (c0, gamma) all floats (note A is always 0 for the h stream)

SPSAOptimizers.QiskitInterface — Module

Using qiskit, one might do:

```

import qiskit_algorithms
import qiskit_algorithms.optimizers as qopt

qiskit_algorithms.utils.algorithm_globals.random_seed = seed

get_eta = lambda: qopt.spsa.powerseries(eta=a0, power=alpha, offset=A)
get_eps = lambda: qopt.spsa.powerseries(eta=c0, power=gamma)

spsa = qopt.SPSA(
    learning_rate = get_eta,
    perturbation = get_eps,
    **kwargs
)

result = spsa.minimize(fn, x0)

```

With this module, one could do this:

```
import SPSAOptimizers.QiskitInterface as QI

QI.seed!(seed)

get_eta = QI.powerseries(; eta=a0, power=alpha, offset=A)
get_eps = QI.powerseries(; eta=c0, power=gamma)

spsa = QI.SPSA(length(x0);
    learning_rate = get_eta,
    perturbation = get_eps,
    kwargs...
)

result = QI.minimize(spsa, fn, x0)

```

Note that QI.SPSA needs a positional argument, not needed in Python. To make up for it, while the result object matches qiskit's interface exactly, the spsa object contains attributes record and trace, which, after the call to QI.minimize, have some extra information unavailable to qiskit (without writing your own callback function).

To avoid expensive and slightly arbitrary calibration steps, this module treats learning_rate and perturbation as mandatory. Similarly, when the kwarg blocking is set to true, the kwarg allowed_increase also becomes mandatory. While the qiskit implementation will perform a calibration to select problem-informed defaults, the calibration is not necessarily ... good. So you really should perform that calibration ahead of time and then manually specify your parameters.

Interesting note: qiskit's initial_hessian kwarg seems to be quite useless. That is, it sets the initial hessian, but the initial hessian is, if I have read the code correctly, always multiplied by zero before it is ever used. So. Don't use it. If you really want to, I've added an extra kwarg initial_hessian_weight to the QI.SPSA interface, which can be a non-negative integer. For the record, the SPSAOptimizers package does permit an initial hessian, but it would requires an extra hyperparameter for an "initial weight", so let's just not use it with this QiskitInterface.

SPSAOptimizers.QiskitInterface.SPSA — Method

SPSA(L; kwargs...)

Initialize an SPSA optimization. The interface is identical to qiskit's, except that there is one required argument L giving the number of parameters, and the learning_rate and perturbation arguments are mandatory.

(Unlike in qiskit, the number of parameters has to be fixed ahead of time to ensure the serializability of the random stream.)

See here for qiskit's interface: https://docs.quantum.ibm.com/api/qiskit/0.28/qiskit.algorithms.optimizers.SPSA

SPSAOptimizers.QiskitInterface.minimize — Method

minimize(spsa, fn, x0)

Run an optimization.

Parameters

spsa - the result of QiskitInterface.SPSA

fn - the function, fn(x) (where x is a vector of parameters)

x0 - the initial guess